rename the Tabs file
