+++ frontend/src/pages/yugioh/SellManagement.tsx
@@ -9,7 +9,7 @@ import VisibilityIcon from '@mui/icons-material/Visibility';
 import PageHeader from '../../components/PageHeader';
 import LensIcon from '@mui/icons-material/Lens';
 import AddIcon from '@mui/icons-material/Add';
-import CardflowTabs from '../../components/sellListing/ListingTopBar';
+import CardflowTabs from '../../components/sellListing/CardflowTabs';
 import { errorToast } from '../../util/errorToast';
 import toast from 'react-hot-toast';
 import { toastMessages } from '../../constants/toast';
